Add simple confirmation on deletion of tag, note, file.

+++ frontend/src/components/tag/tag-list.component.jsx
@@ -23,6 +23,9 @@ const TagList = () => {
     }, [accessToken]);
 
     const handleDelete = async (tag) => {
+        if (!window.confirm(`You are about to delete '${tag.name}' tag that may be associated with notes. Notes will be untagged but not deleted. Do you want to proceed?`)) {
+            return;
+        }
         setTagBeingDeleted(tag);
         await deleteTag(tag._id);
         setTagBeingDeleted({});
