+++ frontend/src/pages/private/note.page.jsx
@@ -19,6 +19,9 @@ const NotePage = ({ match, history }) => {
     const [deleteLoading, setDeleteLoading] = useState(false);
 
     const handleDelete = async () => {
+        if (!window.confirm(`You are about to delete '${note.title}' note. Do you want to proceed?`)) {
+            return;
+        }
         setDeleteLoading(true);
         const isNoteDeleted = await deleteNote(note._id);
         if (isNoteDeleted) {
